# sdks/sdkwork-generations-app-sdk/openapi/sdkwork-generations-app-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SDKWork Generations App API",
    "version": "0.1.0",
    "x-sdkwork-owner": "sdkwork-generations",
    "x-sdkwork-api-authority": "sdkwork-generations-app-api"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
    "/app/v3/api/generations": {
      "get": {
        "tags": [
          "generationsApi"
        ],
        "operationId": "generations.list",
        "x-sdkwork-owner": "sdkwork-generations",
        "x-sdkwork-api-authority": "sdkwork-generations-app-api",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GenerationStatus"
            }
          },
          {
            "name": "modality",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GenerationModality"
            }
          },
          {
            "name": "operation_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Generation record page.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "additionalProperties": false,
                          "required": [
                            "items",
                            "pageInfo"
                          ],
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/GenerationRecord"
                              }
                            },
                            "pageInfo": {
                              "type": "object",
                              "additionalProperties": false,
                              "required": [
                                "mode",
                                "hasMore"
                              ],
                              "properties": {
                                "mode": {
                                  "type": "string",
                                  "enum": [
                                    "cursor"
                                  ]
                                },
                                "nextCursor": {
                                  "type": [
                                    "string",
                                    "null"
                                  ]
                                },
                                "hasMore": {
                                  "type": "boolean"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Problem detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Problem detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/app/v3/api/generations/images/text_to_image": {
      "post": {
        "tags": [
          "generationsApi"
        ],
        "operationId": "generations.images.textToImage",
        "x-sdkwork-owner": "sdkwork-generations",
        "x-sdkwork-api-authority": "sdkwork-generations-app-api",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "description": "Client-supplied idempotency key scoped to the tenant."
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/GenerationCommand"
        },
        "responses": {
          "202": {
            "description": "Generation accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/GenerationCommandResponse"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Problem detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Problem detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Problem detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-idempotent": true
      }
    },
    "/app/v3/api/generations/images/image_edit": {
      "post": {
        "tags": [
          "generationsApi"
        ],
        "operationId": "generations.images.imageEdit",
        "x-sdkwork-owner": "sdkwork-generations",
        "x-sdkwork-api-authority": "sdkwork-generations-app-api",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "description": "Client-supplied idempotency key scoped to the tenant."
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/GenerationCommand"
        },
        "responses": {
          "202": {
            "description": "Generation accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/GenerationCommandResponse"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Problem detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Problem detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Problem detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-idempotent": true
      }
    },
    "/app/v3/api/generations/videos/text_to_video": {
      "post": {
        "tags": [
          "generationsApi"
        ],
        "operationId": "generations.videos.textToVideo",
        "x-sdkwork-owner": "sdkwork-generations",
        "x-sdkwork-api-authority": "sdkwork-generations-app-api",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            },
            "description": "Client-supplied idempotency key scoped to the tenant."
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/GenerationCommand"
        },
        "responses": {
          "202": {
            "description": "Generation accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/GenerationCommandResponse"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Problem detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Problem detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Problem detail.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-idempotent": true
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    },
...
    "requestBodies": {
      "GenerationCommand": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CreateGenerationCommandRequest"
            }
          }
        }
      },
...
    },
...
    "schemas": {
      "GenerationModality": {
        "type": "string",
        "enum": [
          "image",
          "video",
          "music",
          "audio",
          "sfx",
          "voice"
        ]
      },
      "GenerationStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "requires_action",
          "succeeded",
          "failed",
          "canceled"
        ]
      },
...
      "GenerationRecord": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "tenantId",
          "userId",
          "modality",
          "operationType",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "modality": {
            "$ref": "#/components/schemas/GenerationModality"
          },
          "operationType": {
            "type": "string"
          },
          "sourceProvider": {
            "type": "string"
          },
          "sourceJobId": {
            "type": "string"
          },
          "promptPreview": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/GenerationStatus"
          },
          "favorite": {
            "type": "boolean"
          },
          "resultCount": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
...
    }
  }
}